create reusable Allowance Tooltip
